# pkg/trace: replace gogo protobuf with vtprotobuf