--- a/comp/dogstatsd/replay/writer.go
+++ b/comp/dogstatsd/replay/writer.go
@@ -22,9 +22,9 @@ import (
 
 	"github.com/DataDog/datadog-agent/comp/dogstatsd/packets"
 	pb "github.com/DataDog/datadog-agent/pkg/proto/pbgo"
-	"github.com/DataDog/datadog-agent/pkg/proto/utils"
 	"github.com/DataDog/datadog-agent/pkg/tagger"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
+	protoutils "github.com/DataDog/datadog-agent/pkg/util/proto"
 
 	"github.com/golang/protobuf/proto"
 )
@@ -331,7 +331,7 @@ func (tc *TrafficCaptureWriter) writeState() (int, error) {
 			continue
 		}
 
-		entityID, err := utils.Tagger2PbEntityID(entity.ID)
+		entityID, err := protoutils.Tagger2PbEntityID(entity.ID)
 		if err != nil {
 			log.Warnf("unable to compute valid EntityID for %v", id)
 			continue