--- a/internal/tools/proto/tools.go
+++ b/internal/tools/proto/tools.go
@@ -16,6 +16,7 @@ package proto
 //       conflicts with the pins set here.
 
 import (
+	_ "github.com/favadi/protoc-go-inject-tag"
 	_ "github.com/golang/mock/mockgen"
 	_ "github.com/golang/protobuf/protoc-gen-go"
 	_ "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway"