--- a/pkg/proto/go.mod
+++ b/pkg/proto/go.mod
@@ -1,0 +1,20 @@
+module github.com/DataDog/datadog-agent/pkg/proto
+
+go 1.19
+
+require (
+	github.com/golang/mock v1.6.0
+	github.com/golang/protobuf v1.5.3
+	github.com/grpc-ecosystem/grpc-gateway v1.16.0
+	github.com/tinylib/msgp v1.1.8
+	google.golang.org/genproto v0.0.0-20230410155749-daa745c078e1
+	google.golang.org/grpc v1.55.0
+	google.golang.org/protobuf v1.30.0
+)
+
+require (
+	github.com/philhofer/fwd v1.1.2 // indirect
+	golang.org/x/net v0.8.0 // indirect
+	golang.org/x/sys v0.6.0 // indirect
+	golang.org/x/text v0.8.0 // indirect
+)