--- a/pkg/trace/go.mod
+++ b/pkg/trace/go.mod
@@ -11,7 +11,7 @@ replace github.com/docker/distribution => github.com/docker/distribution v2.8.1+
 
 require (
 	github.com/DataDog/datadog-agent/pkg/obfuscate v0.45.0-rc.3
-	github.com/DataDog/datadog-agent/pkg/proto v0.46.0-20230513-devel
+	github.com/DataDog/datadog-agent/pkg/proto v0.46.0-20230514-devel
 	github.com/DataDog/datadog-agent/pkg/remoteconfig/state v0.45.0-rc.3
 	github.com/DataDog/datadog-agent/pkg/util/cgroups v0.45.0-rc.3
 	github.com/DataDog/datadog-agent/pkg/util/log v0.45.0-rc.3