--- a/pkg/trace/pb/agent_payload.proto
+++ b/pkg/trace/pb/agent_payload.proto
@@ -3,6 +3,7 @@
 syntax = "proto3";
 
 package pb;
+option go_package = "github.com/DataDog/datadog-agent/pkg/trace/pb";
 
 import "tracer_payload.proto";