… LICENSE-3rdparty.csv
… cmd/trace-agent/test/backend.go
… cmd/trace-agent/test/testsuite/cards_test.go
… cmd/trace-agent/test/testsuite/chunking_test.go
… cmd/trace-agent/test/testsuite/events_test.go
… cmd/trace-agent/test/testsuite/hostname_test.go
… cmd/trace-agent/test/testsuite/otlp_test.go
… cmd/trace-agent/test/testsuite/proxy_test.go
… cmd/trace-agent/test/testsuite/stats_test.go
… cmd/trace-agent/test/testsuite/testdata/clientstats.go
… cmd/trace-agent/test/testsuite/traces_test.go
… go.sum
… internal/tools/proto/go.sum
… pkg/proto/go.sum
… pkg/proto/pbgo/api.pb.go
… pkg/proto/pbgo/mocks/api_mockgen.pb.go
… pkg/proto/pbgo/model.pb.go
… pkg/proto/pbgo/process.pb.go
… pkg/proto/pbgo/workloadmeta.pb.go
… pkg/trace/agent/agent_test.go
… pkg/trace/agent/fuzz_test.go
… pkg/trace/agent/obfuscate_test.go
… pkg/trace/api/otlp_test.go
… pkg/trace/go.sum
… pkg/trace/pb/agent_payload.pb.go
… pkg/trace/pb/agent_payload_vtproto.pb.go
… pkg/trace/pb/span.pb.go
… pkg/trace/pb/span.proto
… pkg/trace/pb/span_vtproto.pb.go
… pkg/trace/pb/tracer_payload.pb.go