### What does this PR do?
This commit replaces gogo protobuf with the first-party protobuf generator for Go along with a plugin called vtproto (https://github.com/planetscale/vtprotobuf), and regenerates the protocol buffer code.

### Motivation
Gogo is no longer maintained, so we should move off of it. Also, we want to migrate our proto files into the agent's main proto package along with all the other protobuf stuff, which will require us to use the same code generator as the rest of the agent

### Additional Notes
Currently, this code requires 2 third-party generators in addition to the main protobuf generator.

The first, https://github.com/favadi/protoc-go-inject-tag,  is required to add msgpack struct tags to the generated protobuf structs.

The second, https://github.com/planetscale/vtprotobuf, is required for generating efficient, non-reflection-based protobuf marshaling code. The reflection-based marshaling that protoc-gen-go and the proto library do by default are painfully inefficient, and increase CPU usage by a huge amount. vtprotobuf seems to prevent this by providing highly efficient marshaling methods.

The commands to generate the code are in `generate.sh`. 
These should maybe live in `go:generate`, however `go:generate` currently contains msgpack generation commands which clobber our hand-written modifications to the msgpack encoders.

Seeing how this is all going to be moved out of here and into `pkg/proto`, I don't think it's worth refactoring the `go:generate` directives now.


#### Performance
One test just marshaling spans demonstrates how much performance we gain here:
```
### Before
$ go test -v -run XXX -bench BenchmarkSpanProto -benchmem -benchtime 10s .
goos: linux
goarch: amd64
pkg: github.com/DataDog/datadog-agent/pkg/trace/writer
cpu: 11th Gen Intel(R) Core(TM) i9-11950H @ 2.60GHz
BenchmarkSpanProto
BenchmarkSpanProto-16    	 5952206	      2053 ns/op	     848 B/op	      35 allocs/op
PASS
ok  	github.com/DataDog/datadog-agent/pkg/trace/writer	14.260s
```
```
### After
$ go test -v -run XXX -bench BenchmarkSpanProto -benchmem -benchtime 10s .
goos: linux
goarch: amd64
pkg: github.com/DataDog/datadog-agent/pkg/trace/writer
cpu: 11th Gen Intel(R) Core(TM) i9-11950H @ 2.60GHz
BenchmarkSpanProto
BenchmarkSpanProto-16    	33397891	       404.9 ns/op	     208 B/op	       1 allocs/op
PASS
ok  	github.com/DataDog/datadog-agent/pkg/trace/writer	13.892s
```

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes
This change completely rewrites the protobuf types we use both for ingesting msgpack payloads and for sending to intake, so this needs to be reviewed and tested carefully. 

- [ ] Try generating the protobuf
- [ ] Try running a local application with non-trivial traces (traces with many tags, sql queries, children, etc.) and ensure they appear correctly in the backend

In addition to basic local testing, we should monitor deployments to staging to ensure we don't see any disruptions. Dogfooding may catch corner cases we have blind spots for.

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).

Note: Adding GitHub labels is only possible for contributors with write access.
-->

- [ ] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.
- [ ] Use the `major_change` label if your change either has a major impact on the code base, is impacting multiple teams or is changing important well-established internals of the Agent. This label will be use during QA to make sure each team pay extra attention to the changed behavior. For any customer facing change use a releasenote.
- [ ] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [ ] Changed code has automated tests for its functionality.
- [ ] Adequate QA/testing plan information is provided if the `qa/skip-qa` label is not applied.
- [ ] At least one `team/..` label has been applied, indicating the team(s) that should QA this change.
- [ ] If applicable, docs team has been notified or [an issue has been opened on the documentation repo](https://github.com/DataDog/documentation/issues/new).
- [ ] If applicable, the `need-change/operator` and `need-change/helm` labels have been applied.
- [ ] If applicable, the `k8s/<min-version>` label, indicating the lowest Kubernetes version compatible with this feature. 
- [ ] If applicable, the [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated.